Implement feedback (#8)
* Update tests in test/ with given feedback

* Implement PEP 8 standards

* Change class name

* Update '.travis.yml' file
