+++ .travis.yml
@@ -8,7 +8,7 @@ env:
 install:
   - pip install -r requirements.txt
 # command to run tests
-script: nosetests test/tests.py --with-coverage --cover-package=test && coverage report
+script: nosetests test/*.py --with-coverage --cover-package=test && coverage report
 
 after_success: coveralls
